requirements.txt
python-telegram-bot==20.3
SQLAlchemy==1.4.46
aiohttp==3.8.4